requirements.txt
Flask>=2.0
Flask-Session>=0.4.0
APScheduler>=3.6.3
python-dotenv>=0.19.0
requests>=2.25.1
Jinja2>=3.0
sqlite3-bundled-if-needed # sqlite is part of stdlib; keep for clarity